moved invalid chunk with other exceptions
